requirements.txt
python-dotenv
jupyterlab
groq
PyPDF2
python-telegram-bot==22.4
